third_party/WebKit/LayoutTests/svg/dom/svglist-exception-on-out-bounds-error-expected.txt

Tests that out of bounds accesses of SVGList correctly throw exceptions

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgList.getItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Type error.
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.replaceItem(null, index) threw exception TypeError: Type error.
PASS svgList.removeItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.getItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Type error.
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.replaceItem(null, index) threw exception TypeError: Type error.
PASS svgList.removeItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.getItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Type error.
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.replaceItem(null, index) threw exception TypeError: Type error.
PASS svgList.removeItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.getItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Type error.
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.replaceItem(null, index) threw exception TypeError: Type error.
PASS svgList.removeItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.getItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Type error.
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.replaceItem(null, index) threw exception TypeError: Type error.
PASS svgList.removeItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.getItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.insertItemBefore(null, index) threw exception TypeError: Type error.
PASS svgList.insertItemBefore(seg, index) is seg
PASS svgList.replaceItem(seg, index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgList.replaceItem(null, index) threw exception TypeError: Type error.
PASS svgList.removeItem(index) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS successfullyParsed is true

TEST COMPLETE
